+++ src/Router/Dispatcher.php
@@ -86,7 +86,7 @@ class Dispatcher
         $urlData = $this->createDispatcher()
             ->dispatch(strtoupper($method), $path);
 
-        return new DispatchResult($urlData, $this->collector->getCollectedRoutes());
+        return new DispatchResult($urlData, $this->collector);
     }
 
     /**